requirements.txt
pymongo
pymongo[srv]
validators
python-dotenv
boto3
tweepy
beautifulsoup4
Pillow
waitress
flask
selenium
webdriver-manager
flask-cors
tinycss
requests
ngram
lxml
urllib3
scikit-learn
pandas
numpy
langdetect
